requirements.txt
aiogram==2.25.1
sqlalchemy==1.4.41
aiosqlite==0.19.0
environs==9.5.0
google-generativeai==0.3.1
greenlet==2.0.2 